bsps/stm32h7: update STM32 H7 HAL
This patch updates STM32 H7 HAL source files. The files are taken from two
STM projects from their github.com repositories:

(i)
https://github.com/STMicroelectronics/stm32h7xx_hal_driver.git

The project files are still available under BSD-3 license
and the version/commit used is:

d5fc8d05fc16fa2a2a2f948cf6c6ab39e78358e1

which represents post Release v1.11.1 development tree.

(ii)
https://github.com/STMicroelectronics/cmsis_device_h7.git

The project files were re-licensed from previous BSD-3 to Apache 2.0
license. Fortunately the project does not contain NOTICE file so no need
to do anything special when used in RTEMS.

The project version/commit imported is:

6d5ef249bec5177e0e2a0880ed62df2132874d99

which is code-wise Release v1.10.3 exactly.

+++ bsps/arm/stm32h7/include/stm32h7xx_hal_usart_ex.h
@@ -6,13 +6,12 @@
   ******************************************************************************
   * @attention
   *
-  * <h2><center>&copy; Copyright (c) 2017 STMicroelectronics.
-  * All rights reserved.</center></h2>
+  * Copyright (c) 2017 STMicroelectronics.
+  * All rights reserved.
   *
-  * This software component is licensed by ST under BSD 3-Clause license,
-  * the "License"; You may not use this file except in compliance with the
-  * License. You may obtain a copy of the License at:
-  *                        opensource.org/licenses/BSD-3-Clause
+  * This software is licensed under terms that can be found in the LICENSE file
+  * in the root directory of this software component.
+  * If no LICENSE file comes with this software, it is provided AS-IS.
   *
   ******************************************************************************
   */
@@ -39,23 +38,20 @@ extern "C" {
 /* Exported types ------------------------------------------------------------*/
 /* Exported constants --------------------------------------------------------*/
 /** @defgroup USARTEx_Exported_Constants USARTEx Exported Constants
-  * @ingroup RTEMSBSPsARMSTM32H7
   * @{
   */
 
 /** @defgroup USARTEx_Word_Length USARTEx Word Length
-  * @ingroup RTEMSBSPsARMSTM32H7
   * @{
   */
-#define USART_WORDLENGTH_7B                  ((uint32_t)USART_CR1_M1)   /*!< 7-bit long USART frame */
+#define USART_WORDLENGTH_7B                  (USART_CR1_M1)   /*!< 7-bit long USART frame */
 #define USART_WORDLENGTH_8B                  (0x00000000U)              /*!< 8-bit long USART frame */
-#define USART_WORDLENGTH_9B                  ((uint32_t)USART_CR1_M0)   /*!< 9-bit long USART frame */
+#define USART_WORDLENGTH_9B                  (USART_CR1_M0)   /*!< 9-bit long USART frame */
 /**
   * @}
   */
 
 /** @defgroup USARTEx_Slave_Select_management USARTEx Slave Select Management
-  * @ingroup RTEMSBSPsARMSTM32H7
   * @{
   */
 #define USART_NSS_HARD                        0x00000000U          /*!< SPI slave selection depends on NSS input pin              */
@@ -66,7 +62,6 @@ extern "C" {
 
 
 /** @defgroup USARTEx_Slave_Mode USARTEx Synchronous Slave mode enable
-  * @ingroup RTEMSBSPsARMSTM32H7
   * @brief    USART SLAVE mode
   * @{
   */
@@ -77,7 +72,6 @@ extern "C" {
   */
 
 /** @defgroup USARTEx_FIFO_mode USARTEx FIFO  mode
-  * @ingroup RTEMSBSPsARMSTM32H7
   * @brief    USART FIFO  mode
   * @{
   */
@@ -88,7 +82,6 @@ extern "C" {
   */
 
 /** @defgroup USARTEx_TXFIFO_threshold_level USARTEx TXFIFO threshold level
-  * @ingroup RTEMSBSPsARMSTM32H7
   * @brief    USART TXFIFO level
   * @{
   */
@@ -103,7 +96,6 @@ extern "C" {
   */
 
 /** @defgroup USARTEx_RXFIFO_threshold_level USARTEx RXFIFO threshold level
-  * @ingroup RTEMSBSPsARMSTM32H7
   * @brief    USART RXFIFO level
   * @{
   */
@@ -123,7 +115,6 @@ extern "C" {
 
 /* Private macros ------------------------------------------------------------*/
 /** @defgroup USARTEx_Private_Macros USARTEx Private Macros
-  * @ingroup RTEMSBSPsARMSTM32H7
   * @{
   */
 
@@ -289,4 +280,3 @@ HAL_StatusTypeDef HAL_USARTEx_SetRxFifoThreshold(USART_HandleTypeDef *husart, ui
 
 #endif /* STM32H7xx_HAL_USART_EX_H */
 
-/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
